feat(c-api) Move `memory` into its own module.
