+++ lib/c-api/src/wasm_c_api/externals/mod.rs
@@ -1,1 +1,2 @@
 pub mod global;
+pub mod memory;
